PR S4.2-followup: extract views/xp/ and cogs/xp/_helpers.py
Completes the F-3 decomposition convention for the XP subsystem so the
cog file is back to a thin Discord adapter (110 LOC, was 601) and every
UI surface lives where readers expect it.

New layout (mirrors views/economy/, views/mining/, cogs/economy/):

  cogs/xp_cog.py          — commands + setup + listener wiring only
  cogs/xp/__init__.py     — package docstring updated
  cogs/xp/listener.py     — unchanged (S4.2)
  cogs/xp/_helpers.py     — _STAT_TYPES, _guild_xp_settings,
                            _progress_bar, _build_rank_embed
  views/xp/__init__.py    — re-exports
  views/xp/main_panel.py  — _XpHubView (HubView)
  views/xp/rank_view.py   — _RankView + _RankSelect (!rank dropdown)
  views/xp/config_panel.py — XpConfigView (!xpconfig)
  views/xp/modals.py      — 5 modals (give/reset/range/cooldown/channel)

Behavior preserved end-to-end:

  * All command flows (!xpmenu, !rank, !givexp, !resetxp, !xpconfig)
    route through the same services / db calls / events.
  * HubView → modal spawning uses local imports inside button
    callbacks to break the natural xp_cog ↔ views.xp cycle.
  * XpConfigView modals still invoke invalidate_xp_config so the F-1
    cache picks up new XP_MIN / XP_MAX / XP_COOLDOWN /
    XP_ANNOUNCE_CHANNEL on the next on_message read.
  * build_help_menu_view hook still returns the hub panel for the
    help-cog direct-navigation path.

Tests updated, not weakened:

  * test_xp_cog_admin_routes.py — modal-level patches now target
    views.xp.modals.{_parse_member, xp_service.award, xp_service.reset};
    command-level patches keep cogs.xp_cog.xp_service.* as the cog
    still owns those mutation entry points.
  * test_xp_cog_caching.py — the static invariant assertion moved
    from cogs.xp_cog to views.xp.modals (where the 3 admin XP-config
    mutation paths now live) with an updated rationale docstring.

Dead code removed: _XP_MENU_COMMANDS, _XP_MIN, _XP_MAX, _COOLDOWN —
all module-level constants left behind by the F-1 cache migration
and unreferenced by any caller (verified via grep).

Validation:

  * pytest tests/unit/ → 837 passed, 0 failed
  * INV-A..INV-N (incl. INV-G xp service ownership) green
  * Identity contract STRICT mode green (5 surfaces consistent)
  * black, isort, ruff, mypy all clean over the whole disbot/ tree

This is the bottleneck stage for Phase A: A2 (moderation), A3 (rps),
A4 (blackjack), A5 (diagnostic) all reuse this F-3 layout as their
template, and A6 (soft size invariant test) gates on every cog being
≤ 800 LOC, which xp_cog now is (110).
