+++ disbot/cogs/xp/__init__.py
@@ -1,4 +1,4 @@
-"""XP subsystem — domain logic (S4.2).
+"""XP subsystem — domain logic (S4.2 + S4.2-followup).
 
 See ``docs/architecture.md`` §"Subsystem decomposition" for the layout
 convention.  Modules in this package are pure-ish: they call DB +
@@ -7,4 +7,6 @@ listener body itself.
 
 Modules:
     listener  — on_message handler + level-up announcement + role assignment
+    _helpers  — rank embed builder, progress bar, cached config tuple shim
+                (shared by ``cogs/xp_cog.py`` and ``views/xp/*``)
 """
